rename components, modify css
