+++ src/components/CopyButton.tsx
@@ -1,4 +1,4 @@
-import { SvgIcon } from '@/components/SvgIcon';
+import { Icon } from '@/components/Icon';
 
 import '@/components/copy-button.css';
 
@@ -11,7 +11,7 @@ export function CopyButton({
 }) {
   return (
     <button className="button copy-to-clipboard-button" value={value}>
-      <SvgIcon icon="copy" />
+      <Icon icon="copy" />
       {label ? <span className="copy-button-label">{label}</span> : null}
     </button>
   );
