+++ src/components/contact-section.css
@@ -25,7 +25,7 @@
   box-sizing: border-box;
   cursor: pointer;
   display: flex;
-  font-family: monospace;
+  /* font-family: monospace; */
   height: var(--padding-x2);
   justify-content: center;
   padding: 0 1.125rem;
