+++ src/components/home-section.css
@@ -2,10 +2,18 @@
   align-items: center;
   display: flex;
   flex-direction: column;
+  grid-area: home;
+  padding-top: var(--header-height);
+}
+@media screen and (min-width: 600px) {
+  .home-section {
+    /* position: fixed; */
+  }
 }
 
 .name {
   font-size: clamp(2.5rem, 8vw, 5rem);
+  font-size: 2.5rem;
   font-weight: 300;
   line-height: 1.6;
   margin: 0;
@@ -15,6 +23,7 @@
 
 .job-title {
   font-size: clamp(2rem, 6vw, 4rem);
+  font-size: 2rem;
   font-weight: 300;
   line-height: 1.6;
   margin: 8px 0 0 0;
