+++ src/components/projects-section.css
@@ -7,8 +7,6 @@
   display: grid;
   gap: 1rem;
   grid-template-columns: 1fr;
-  margin-top: 1rem;
-  /* max-width: 1200px; */
 }
 
 .project {
@@ -55,8 +53,8 @@
   }
 }
 
-@media screen and (min-width: 900px) {
+/* @media screen and (min-width: 900px) {
   .projects-grid {
     grid-template-columns: repeat(3, 1fr);
   }
-}
+} */
